Cleanup ServerRequest tests.
Remove duplicate tests and move tests using ServerRequestFactory to it's test file.
