+++ tests/TestCase/Http/ServerRequestFactoryTest.php
@@ -125,6 +125,165 @@ class ServerRequestFactoryTest extends TestCase
         $this->assertSame('/urlencode%20me', $res->getAttribute('base'));
         $this->assertSame('/urlencode%20me/', $res->getAttribute('webroot'));
         $this->assertSame('/posts/view/1', $res->getUri()->getPath());
+
+        $request = ServerRequestFactory::fromGlobals([
+            'DOCUMENT_ROOT' => '/cake/repo/branches',
+            'PHP_SELF' => '/1.2.x.x/webroot/index.php',
+            'PATH_INFO' => '/posts/view/1',
+        ]);
+        $this->assertSame('/1.2.x.x', $request->getAttribute('base'));
+        $this->assertSame('/1.2.x.x/', $request->getAttribute('webroot'));
+        $this->assertSame('/posts/view/1', $request->getRequestTarget());
+
+        $request = ServerRequestFactory::fromGlobals([
+            'DOCUMENT_ROOT' => '/cake/repo/branches/1.2.x.x/test/',
+            'PHP_SELF' => '/webroot/index.php',
+        ]);
+
+        $this->assertSame('', $request->getAttribute('base'));
+        $this->assertSame('/', $request->getAttribute('webroot'));
+
+        $request = ServerRequestFactory::fromGlobals([
+            'DOCUMENT_ROOT' => '/some/apps/where',
+            'PHP_SELF' => '/webroot/index.php',
+        ]);
+
+        $this->assertSame('', $request->getAttribute('base'));
+        $this->assertSame('/', $request->getAttribute('webroot'));
+
+        Configure::write('App.dir', 'auth');
+
+        $request = ServerRequestFactory::fromGlobals([
+            'DOCUMENT_ROOT' => '/cake/repo/branches',
+            'PHP_SELF' => '/demos/webroot/index.php',
+        ]);
+
+        $this->assertSame('/demos', $request->getAttribute('base'));
+        $this->assertSame('/demos/', $request->getAttribute('webroot'));
+
+        Configure::write('App.dir', 'code');
+
+        $request = ServerRequestFactory::fromGlobals([
+            'DOCUMENT_ROOT' => '/Library/WebServer/Documents',
+            'PHP_SELF' => '/clients/PewterReport/webroot/index.php',
+        ]);
+
+        $this->assertSame('/clients/PewterReport', $request->getAttribute('base'));
+        $this->assertSame('/clients/PewterReport/', $request->getAttribute('webroot'));
+    }
+
+    /**
+     * Test baseUrl with ModRewrite alias
+     *
+     * @return void
+     */
+    public function testBaseUrlwithModRewriteAlias()
+    {
+        Configure::write('App.base', '/control');
+
+        $request = ServerRequestFactory::fromGlobals([
+            'DOCUMENT_ROOT' => '/home/aplusnur/public_html',
+            'PHP_SELF' => '/control/index.php',
+        ]);
+
+        $this->assertSame('/control', $request->getAttribute('base'));
+        $this->assertSame('/control/', $request->getAttribute('webroot'));
+
+        Configure::write('App.base', false);
+        Configure::write('App.dir', 'affiliate');
+        Configure::write('App.webroot', 'newaffiliate');
+
+        $request = ServerRequestFactory::fromGlobals([
+            'DOCUMENT_ROOT' => '/var/www/abtravaff/html',
+            'PHP_SELF' => '/newaffiliate/index.php',
+        ]);
+
+        $this->assertSame('', $request->getAttribute('base'));
+        $this->assertSame('/', $request->getAttribute('webroot'));
+    }
+
+    /**
+     * Test base, webroot, URL and here parsing when there is URL rewriting but
+     * CakePHP gets called with index.php in URL nonetheless.
+     *
+     * Tests uri with
+     *
+     * - index.php/
+     * - index.php/
+     * - index.php/apples/
+     * - index.php/bananas/eat/tasty_banana
+     *
+     * @return void
+     */
+    public function testBaseUrlWithModRewriteAndIndexPhp()
+    {
+        $request = ServerRequestFactory::fromGlobals([
+            'DOCUMENT_ROOT' => '/cakephp/webroot/index.php',
+            'PHP_SELF' => '/cakephp/webroot/index.php',
+        ]);
+
+        $this->assertSame('/cakephp', $request->getAttribute('base'));
+        $this->assertSame('/cakephp/', $request->getAttribute('webroot'));
+        $this->assertSame('/', $request->getRequestTarget());
+
+        $request = ServerRequestFactory::fromGlobals([
+            'REQUEST_URI' => '/cakephp/webroot/index.php/',
+            'PHP_SELF' => '/cakephp/webroot/index.php/',
+            'PATH_INFO' => '/',
+        ]);
+
+        $this->assertSame('/cakephp', $request->getAttribute('base'));
+        $this->assertSame('/cakephp/', $request->getAttribute('webroot'));
+        $this->assertSame('/', $request->getRequestTarget());
+
+        $request = ServerRequestFactory::fromGlobals([
+            'REQUEST_URI' => '/cakephp/webroot/index.php/apples',
+            'PHP_SELF' => '/cakephp/webroot/index.php/apples',
+            'PATH_INFO' => '/apples',
+        ]);
+
+        $this->assertSame('/cakephp', $request->getAttribute('base'));
+        $this->assertSame('/cakephp/', $request->getAttribute('webroot'));
+        $this->assertSame('/apples', $request->getRequestTarget());
+
+        $request = ServerRequestFactory::fromGlobals([
+            'REQUEST_URI' => '/cakephp/webroot/index.php/melons/share/',
+            'PHP_SELF' => '/cakephp/webroot/index.php/melons/share/',
+            'PATH_INFO' => '/melons/share/',
+        ]);
+
+        $this->assertSame('/cakephp', $request->getAttribute('base'));
+        $this->assertSame('/cakephp/', $request->getAttribute('webroot'));
+        $this->assertSame('/melons/share/', $request->getRequestTarget());
+
+        $request = ServerRequestFactory::fromGlobals([
+            'REQUEST_URI' => '/cakephp/webroot/index.php/bananas/eat/tasty_banana',
+            'PHP_SELF' => '/cakephp/webroot/index.php/bananas/eat/tasty_banana',
+            'PATH_INFO' => '/bananas/eat/tasty_banana',
+        ]);
+
+        $this->assertSame('/cakephp', $request->getAttribute('base'));
+        $this->assertSame('/cakephp/', $request->getAttribute('webroot'));
+        $this->assertSame('/bananas/eat/tasty_banana', $request->getRequestTarget());
+    }
+
+    /**
+     * Test that even if mod_rewrite is on, and the url contains index.php
+     * and there are numerous //s that the base/webroot is calculated correctly.
+     *
+     * @return void
+     */
+    public function testBaseUrlWithModRewriteAndExtraSlashes()
+    {
+        $request = ServerRequestFactory::fromGlobals([
+            'REQUEST_URI' => '/cakephp/webroot///index.php/bananas/eat',
+            'PHP_SELF' => '/cakephp/webroot///index.php/bananas/eat',
+            'PATH_INFO' => '/bananas/eat',
+        ]);
+
+        $this->assertSame('/cakephp', $request->getAttribute('base'));
+        $this->assertSame('/cakephp/', $request->getAttribute('webroot'));
+        $this->assertSame('/bananas/eat', $request->getRequestTarget());
     }
 
     /**
@@ -225,6 +384,556 @@ class ServerRequestFactoryTest extends TestCase
         $this->assertSame('/webroot/', $res->getAttribute('webroot'));
         $this->assertSame('/index.php', $res->getAttribute('base'));
         $this->assertSame('/posts/add', $res->getUri()->getPath());
+    }
+
+    /**
+     * Check that a sub-directory containing app|webroot doesn't get mishandled when re-writing is off.
+     *
+     * @return void
+     */
+    public function testBaseUrlWithAppAndWebrootInDirname()
+    {
+        Configure::write('App.baseUrl', '/approval/index.php');
+
+        $request = ServerRequestFactory::fromGlobals([
+            'DOCUMENT_ROOT' => '/Users/markstory/Sites/',
+            'SCRIPT_FILENAME' => '/Users/markstory/Sites/approval/index.php',
+        ]);
+        $this->assertSame('/approval/index.php', $request->getAttribute('base'));
+        $this->assertSame('/approval/webroot/', $request->getAttribute('webroot'));
+
+        Configure::write('App.baseUrl', '/webrootable/index.php');
+
+        $request = ServerRequestFactory::fromGlobals([
+            'DOCUMENT_ROOT' => '/Users/markstory/Sites/',
+            'SCRIPT_FILENAME' => '/Users/markstory/Sites/webrootable/index.php',
+        ]);
+        $this->assertSame('/webrootable/index.php', $request->getAttribute('base'));
+        $this->assertSame('/webrootable/webroot/', $request->getAttribute('webroot'));
+    }
+
+    /**
+     * Test baseUrl and webroot with baseUrl
+     *
+     * @return void
+     */
+    public function testBaseUrlAndWebrootWithBaseUrl()
+    {
+        Configure::write('App.dir', 'App');
+        Configure::write('App.baseUrl', '/App/webroot/index.php');
+
+        $request = ServerRequestFactory::fromGlobals();
+        $this->assertSame('/App/webroot/index.php', $request->getAttribute('base'));
+        $this->assertSame('/App/webroot/', $request->getAttribute('webroot'));
+
+        Configure::write('App.baseUrl', '/App/webroot/test.php');
+        $request = ServerRequestFactory::fromGlobals();
+        $this->assertSame('/App/webroot/test.php', $request->getAttribute('base'));
+        $this->assertSame('/App/webroot/', $request->getAttribute('webroot'));
+
+        Configure::write('App.baseUrl', '/App/index.php');
+        $request = ServerRequestFactory::fromGlobals();
+        $this->assertSame('/App/index.php', $request->getAttribute('base'));
+        $this->assertSame('/App/webroot/', $request->getAttribute('webroot'));
+
+        Configure::write('App.baseUrl', '/CakeBB/App/webroot/index.php');
+        $request = ServerRequestFactory::fromGlobals();
+        $this->assertSame('/CakeBB/App/webroot/index.php', $request->getAttribute('base'));
+        $this->assertSame('/CakeBB/App/webroot/', $request->getAttribute('webroot'));
+
+        Configure::write('App.baseUrl', '/CakeBB/App/index.php');
+        $request = ServerRequestFactory::fromGlobals();
+
+        $this->assertSame('/CakeBB/App/index.php', $request->getAttribute('base'));
+        $this->assertSame('/CakeBB/App/webroot/', $request->getAttribute('webroot'));
+
+        Configure::write('App.baseUrl', '/CakeBB/index.php');
+        $request = ServerRequestFactory::fromGlobals();
+
+        $this->assertSame('/CakeBB/index.php', $request->getAttribute('base'));
+        $this->assertSame('/CakeBB/webroot/', $request->getAttribute('webroot'));
+
+        Configure::write('App.baseUrl', '/dbhauser/index.php');
+        $request = ServerRequestFactory::fromGlobals([
+            'DOCUMENT_ROOT' => '/kunden/homepages/4/d181710652/htdocs/joomla',
+            'SCRIPT_FILENAME' => '/kunden/homepages/4/d181710652/htdocs/joomla/dbhauser/index.php',
+        ]);
+
+        $this->assertSame('/dbhauser/index.php', $request->getAttribute('base'));
+        $this->assertSame('/dbhauser/webroot/', $request->getAttribute('webroot'));
+    }
+
+    /**
+     * Test that a request with a . in the main GET parameter is filtered out.
+     * PHP changes GET parameter keys containing dots to _.
+     *
+     * @return void
+     */
+    public function testGetParamsWithDot()
+    {
+        $request = ServerRequestFactory::fromGlobals([
+            'PHP_SELF' => '/webroot/index.php',
+            'REQUEST_URI' => '/posts/index/add.add',
+        ]);
+        $this->assertSame('', $request->getAttribute('base'));
+        $this->assertEquals([], $request->getQueryParams());
+
+        $request = ServerRequestFactory::fromGlobals([
+            'PHP_SELF' => '/cake_dev/webroot/index.php',
+            'REQUEST_URI' => '/cake_dev/posts/index/add.add',
+        ]);
+        $this->assertSame('/cake_dev', $request->getAttribute('base'));
+        $this->assertEquals([], $request->getQueryParams());
+    }
+
+    /**
+     * Test that a request with urlencoded bits in the main GET parameter are filtered out.
+     *
+     * @return void
+     */
+    public function testGetParamWithUrlencodedElement()
+    {
+        $request = ServerRequestFactory::fromGlobals([
+            'PHP_SELF' => '/webroot/index.php',
+            'REQUEST_URI' => '/posts/add/%E2%88%82%E2%88%82',
+        ]);
+        $this->assertSame('', $request->getAttribute('base'));
+        $this->assertEquals([], $request->getQueryParams());
+
+        $request = ServerRequestFactory::fromGlobals([
+            'PHP_SELF' => '/cake_dev/webroot/index.php',
+            'REQUEST_URI' => '/cake_dev/posts/add/%E2%88%82%E2%88%82',
+        ]);
+        $this->assertSame('/cake_dev', $request->getAttribute('base'));
+        $this->assertEquals([], $request->getQueryParams());
+    }
+
+    /**
+     * Generator for environment configurations
+     *
+     * @return array Environment array
+     */
+    public static function environmentGenerator()
+    {
+        return [
+            [
+                'IIS - No rewrite base path',
+                [
+                    'App' => [
+                        'base' => false,
+                        'baseUrl' => '/index.php',
+                        'dir' => 'TestApp',
+                        'webroot' => 'webroot',
+                    ],
+                    'SERVER' => [
+                        'SCRIPT_NAME' => '/index.php',
+                        'PATH_TRANSLATED' => 'C:\\Inetpub\\wwwroot',
+                        'QUERY_STRING' => '',
+                        'REQUEST_URI' => '/index.php',
+                        'URL' => '/index.php',
+                        'SCRIPT_FILENAME' => 'C:\\Inetpub\\wwwroot\\index.php',
+                        'ORIG_PATH_INFO' => '/index.php',
+                        'PATH_INFO' => '',
+                        'ORIG_PATH_TRANSLATED' => 'C:\\Inetpub\\wwwroot\\index.php',
+                        'DOCUMENT_ROOT' => 'C:\\Inetpub\\wwwroot',
+                        'PHP_SELF' => '/index.php',
+                    ],
+                ],
+                [
+                    'base' => '/index.php',
+                    'webroot' => '/webroot/',
+                    'url' => '',
+                ],
+            ],
+            [
+                'IIS - No rewrite with path, no PHP_SELF',
+                [
+                    'App' => [
+                        'base' => false,
+                        'baseUrl' => '/index.php?',
+                        'dir' => 'TestApp',
+                        'webroot' => 'webroot',
+                    ],
+                    'SERVER' => [
+                        'QUERY_STRING' => '/posts/add',
+                        'REQUEST_URI' => '/index.php?/posts/add',
+                        'PHP_SELF' => '',
+                        'URL' => '/index.php?/posts/add',
+                        'DOCUMENT_ROOT' => 'C:\\Inetpub\\wwwroot',
+                        'argv' => ['/posts/add'],
+                        'argc' => 1,
+                    ],
+                ],
+                [
+                    'url' => 'posts/add',
+                    'base' => '/index.php?',
+                    'webroot' => '/webroot/',
+                ],
+            ],
+            [
+                'IIS - No rewrite sub dir 2',
+                [
+                    'App' => [
+                        'base' => false,
+                        'baseUrl' => '/site/index.php',
+                        'dir' => 'TestApp',
+                        'webroot' => 'webroot',
+                    ],
+                    'SERVER' => [
+                        'SCRIPT_NAME' => '/site/index.php',
+                        'PATH_TRANSLATED' => 'C:\\Inetpub\\wwwroot',
+                        'QUERY_STRING' => '',
+                        'REQUEST_URI' => '/site/index.php',
+                        'URL' => '/site/index.php',
+                        'SCRIPT_FILENAME' => 'C:\\Inetpub\\wwwroot\\site\\index.php',
+                        'DOCUMENT_ROOT' => 'C:\\Inetpub\\wwwroot',
+                        'PHP_SELF' => '/site/index.php',
+                        'argv' => [],
+                        'argc' => 0,
+                    ],
+                ],
+                [
+                    'url' => '',
+                    'base' => '/site/index.php',
+                    'webroot' => '/site/webroot/',
+                ],
+            ],
+            [
+                'IIS - No rewrite sub dir 2 with path',
+                [
+                    'App' => [
+                        'base' => false,
+                        'baseUrl' => '/site/index.php',
+                        'dir' => 'TestApp',
+                        'webroot' => 'webroot',
+                    ],
+                    'SERVER' => [
+                        'SCRIPT_NAME' => '/site/index.php',
+                        'PATH_TRANSLATED' => 'C:\\Inetpub\\wwwroot',
+                        'QUERY_STRING' => '/posts/add',
+                        'REQUEST_URI' => '/site/index.php/posts/add',
+                        'URL' => '/site/index.php/posts/add',
+                        'ORIG_PATH_TRANSLATED' => 'C:\\Inetpub\\wwwroot\\site\\index.php',
+                        'DOCUMENT_ROOT' => 'C:\\Inetpub\\wwwroot',
+                        'PHP_SELF' => '/site/index.php/posts/add',
+                        'argv' => ['/posts/add'],
+                        'argc' => 1,
+                    ],
+                ],
+                [
+                    'url' => 'posts/add',
+                    'base' => '/site/index.php',
+                    'webroot' => '/site/webroot/',
+                ],
+            ],
+            [
+                'Apache - No rewrite, document root set to webroot, requesting path',
+                [
+                    'App' => [
+                        'base' => false,
+                        'baseUrl' => '/index.php',
+                        'dir' => 'TestApp',
+                        'webroot' => 'webroot',
+                    ],
+                    'SERVER' => [
+                        'DOCUMENT_ROOT' => '/Library/WebServer/Documents/site/App/webroot',
+                        'SCRIPT_FILENAME' => '/Library/WebServer/Documents/site/App/webroot/index.php',
+                        'QUERY_STRING' => '',
+                        'REQUEST_URI' => '/index.php/posts/index',
+                        'SCRIPT_NAME' => '/index.php',
+                        'PATH_INFO' => '/posts/index',
+                        'PHP_SELF' => '/index.php/posts/index',
+                    ],
+                ],
+                [
+                    'url' => 'posts/index',
+                    'base' => '/index.php',
+                    'webroot' => '/',
+                ],
+            ],
+            [
+                'Apache - No rewrite, document root set to webroot, requesting root',
+                [
+                    'App' => [
+                        'base' => false,
+                        'baseUrl' => '/index.php',
+                        'dir' => 'TestApp',
+                        'webroot' => 'webroot',
+                    ],
+                    'SERVER' => [
+                        'DOCUMENT_ROOT' => '/Library/WebServer/Documents/site/App/webroot',
+                        'SCRIPT_FILENAME' => '/Library/WebServer/Documents/site/App/webroot/index.php',
+                        'QUERY_STRING' => '',
+                        'REQUEST_URI' => '/index.php',
+                        'SCRIPT_NAME' => '/index.php',
+                        'PATH_INFO' => '',
+                        'PHP_SELF' => '/index.php',
+                    ],
+                ],
+                [
+                    'url' => '',
+                    'base' => '/index.php',
+                    'webroot' => '/',
+                ],
+            ],
+            [
+                'Apache - No rewrite, document root set above top level cake dir, requesting path',
+                [
+                    'App' => [
+                        'base' => false,
+                        'baseUrl' => '/site/index.php',
+                        'dir' => 'TestApp',
+                        'webroot' => 'webroot',
+                    ],
+                    'SERVER' => [
+                        'SERVER_NAME' => 'localhost',
+                        'DOCUMENT_ROOT' => '/Library/WebServer/Documents',
+                        'SCRIPT_FILENAME' => '/Library/WebServer/Documents/site/index.php',
+                        'REQUEST_URI' => '/site/index.php/posts/index',
+                        'SCRIPT_NAME' => '/site/index.php',
+                        'PATH_INFO' => '/posts/index',
+                        'PHP_SELF' => '/site/index.php/posts/index',
+                    ],
+                ],
+                [
+                    'url' => 'posts/index',
+                    'base' => '/site/index.php',
+                    'webroot' => '/site/webroot/',
+                ],
+            ],
+            [
+                'Apache - No rewrite, document root set above top level cake dir, request root, no PATH_INFO',
+                [
+                    'App' => [
+                        'base' => false,
+                        'baseUrl' => '/site/index.php',
+                        'dir' => 'TestApp',
+                        'webroot' => 'webroot',
+                    ],
+                    'SERVER' => [
+                        'SERVER_NAME' => 'localhost',
+                        'DOCUMENT_ROOT' => '/Library/WebServer/Documents',
+                        'SCRIPT_FILENAME' => '/Library/WebServer/Documents/site/index.php',
+                        'REQUEST_URI' => '/site/index.php/',
+                        'SCRIPT_NAME' => '/site/index.php',
+                        'PHP_SELF' => '/site/index.php/',
+                    ],
+                ],
+                [
+                    'url' => '',
+                    'base' => '/site/index.php',
+                    'webroot' => '/site/webroot/',
+                ],
+            ],
+            [
+                'Apache - No rewrite, document root set above top level cake dir, request path, with GET',
+                [
+                    'App' => [
+                        'base' => false,
+                        'baseUrl' => '/site/index.php',
+                        'dir' => 'TestApp',
+                        'webroot' => 'webroot',
+                    ],
+                    'GET' => ['a' => 'b', 'c' => 'd'],
+                    'SERVER' => [
+                        'SERVER_NAME' => 'localhost',
+                        'DOCUMENT_ROOT' => '/Library/WebServer/Documents',
+                        'SCRIPT_FILENAME' => '/Library/WebServer/Documents/site/index.php',
+                        'REQUEST_URI' => '/site/index.php/posts/index?a=b&c=d',
+                        'SCRIPT_NAME' => '/site/index.php',
+                        'PATH_INFO' => '/posts/index',
+                        'PHP_SELF' => '/site/index.php/posts/index',
+                        'QUERY_STRING' => 'a=b&c=d',
+                    ],
+                ],
+                [
+                    'urlParams' => ['a' => 'b', 'c' => 'd'],
+                    'url' => 'posts/index',
+                    'base' => '/site/index.php',
+                    'webroot' => '/site/webroot/',
+                ],
+            ],
+            [
+                'Apache - w/rewrite, document root set above top level cake dir, request root, no PATH_INFO',
+                [
+                    'App' => [
+                        'base' => false,
+                        'baseUrl' => false,
+                        'dir' => 'TestApp',
+                        'webroot' => 'webroot',
+                    ],
+                    'SERVER' => [
+                        'SERVER_NAME' => 'localhost',
+                        'DOCUMENT_ROOT' => '/Library/WebServer/Documents',
+                        'SCRIPT_FILENAME' => '/Library/WebServer/Documents/site/index.php',
+                        'REQUEST_URI' => '/site/',
+                        'SCRIPT_NAME' => '/site/webroot/index.php',
+                        'PHP_SELF' => '/site/webroot/index.php',
+                    ],
+                ],
+                [
+                    'url' => '',
+                    'base' => '/site',
+                    'webroot' => '/site/',
+                ],
+            ],
+            [
+                'Apache - w/rewrite, document root above top level cake dir, request root, no PATH_INFO/REQUEST_URI',
+                [
+                    'App' => [
+                        'base' => false,
+                        'baseUrl' => false,
+                        'dir' => 'TestApp',
+                        'webroot' => 'webroot',
+                    ],
+                    'SERVER' => [
+                        'SERVER_NAME' => 'localhost',
+                        'DOCUMENT_ROOT' => '/Library/WebServer/Documents',
+                        'SCRIPT_FILENAME' => '/Library/WebServer/Documents/site/index.php',
+                        'SCRIPT_NAME' => '/site/webroot/index.php',
+                        'PHP_SELF' => '/site/webroot/index.php',
+                        'PATH_INFO' => null,
+                        'REQUEST_URI' => null,
+                    ],
+                ],
+                [
+                    'url' => '',
+                    'base' => '/site',
+                    'webroot' => '/site/',
+                ],
+            ],
+            [
+                'Apache - w/rewrite, document root set to webroot, request root, no PATH_INFO/REQUEST_URI',
+                [
+                    'App' => [
+                        'base' => false,
+                        'baseUrl' => false,
+                        'dir' => 'TestApp',
+                        'webroot' => 'webroot',
+                    ],
+                    'SERVER' => [
+                        'SERVER_NAME' => 'localhost',
+                        'DOCUMENT_ROOT' => '/Library/WebServer/Documents/site/webroot',
+                        'SCRIPT_FILENAME' => '/Library/WebServer/Documents/site/webroot/index.php',
+                        'SCRIPT_NAME' => '/index.php',
+                        'PHP_SELF' => '/index.php',
+                        'PATH_INFO' => null,
+                        'REQUEST_URI' => null,
+                    ],
+                ],
+                [
+                    'url' => '',
+                    'base' => '',
+                    'webroot' => '/',
+                ],
+            ],
+            [
+                'Apache - w/rewrite, document root set above top level cake dir, request root, absolute REQUEST_URI',
+                [
+                    'App' => [
+                        'base' => false,
+                        'baseUrl' => false,
+                        'dir' => 'TestApp',
+                        'webroot' => 'webroot',
+                    ],
+                    'SERVER' => [
+                        'SERVER_NAME' => 'localhost',
+                        'DOCUMENT_ROOT' => '/Library/WebServer/Documents',
+                        'SCRIPT_FILENAME' => '/Library/WebServer/Documents/site/index.php',
+                        'REQUEST_URI' => '/site/posts/index',
+                        'SCRIPT_NAME' => '/site/webroot/index.php',
+                        'PHP_SELF' => '/site/webroot/index.php',
+                    ],
+                ],
+                [
+                    'url' => 'posts/index',
+                    'base' => '/site',
+                    'webroot' => '/site/',
+                ],
+            ],
+            [
+                'Nginx - w/rewrite, document root set to webroot, request root, no PATH_INFO',
+                [
+                    'App' => [
+                        'base' => false,
+                        'baseUrl' => false,
+                        'dir' => 'TestApp',
+                        'webroot' => 'webroot',
+                    ],
+                    'SERVER' => [
+                        'SERVER_NAME' => 'localhost',
+                        'DOCUMENT_ROOT' => '/Library/WebServer/Documents/site/webroot',
+                        'SCRIPT_FILENAME' => '/Library/WebServer/Documents/site/webroot/index.php',
+                        'SCRIPT_NAME' => '/index.php',
+                        'PHP_SELF' => '/index.php',
+                        'PATH_INFO' => null,
+                        'REQUEST_URI' => '/posts/add',
+                    ],
+                ],
+                [
+                    'url' => 'posts/add',
+                    'base' => '',
+                    'webroot' => '/',
+                    'urlParams' => [],
+                ],
+            ],
+            [
+                'Nginx - w/rewrite, document root set above top level cake dir, request root, no PATH_INFO, base parameter set',
+                [
+                    'App' => [
+                        'base' => false,
+                        'baseUrl' => false,
+                        'dir' => 'app',
+                        'webroot' => 'webroot',
+                    ],
+                    'SERVER' => [
+                        'SERVER_NAME' => 'localhost',
+                        'DOCUMENT_ROOT' => '/Library/WebServer/Documents',
+                        'SCRIPT_FILENAME' => '/Library/WebServer/Documents/site/App/webroot/index.php',
+                        'SCRIPT_NAME' => '/site/app/webroot/index.php',
+                        'PHP_SELF' => '/site/webroot/index.php',
+                        'PATH_INFO' => null,
+                        'REQUEST_URI' => '/site/posts/add',
+                    ],
+                ],
+                [
+                    'url' => 'posts/add',
+                    'base' => '/site',
+                    'webroot' => '/site/',
+                    'urlParams' => [],
+                ],
+            ],
+        ];
+    }
+
+    /**
+     * Test environment detection
+     *
+     * @dataProvider environmentGenerator
+     * @param $name
+     * @param $data
+     * @param $expected
+     * @return void
+     */
+    public function testEnvironmentDetection($name, $data, $expected)
+    {
+        if (isset($data['App'])) {
+            Configure::write('App', $data['App']);
+        }
+
+        $request = ServerRequestFactory::fromGlobals(
+            $data['SERVER'] ?? null,
+            $data['GET'] ?? null
+        );
+        $uri = $request->getUri();
+
+        $this->assertSame('/' . $expected['url'], $uri->getPath(), 'Uri->getPath() is incorrect');
+        $this->assertEquals($expected['base'], $request->getAttribute('base'), 'base is incorrect');
+        $this->assertEquals($expected['webroot'], $request->getAttribute('webroot'), 'webroot is incorrect');
+
+        if (isset($expected['urlParams'])) {
+            $this->assertEquals($expected['urlParams'], $request->getQueryParams(), 'GET param mismatch');
+        }
     }
 
     /**
